+++ includes/post.php
@@ -15,5 +15,5 @@
  
  // Review
  if (Input::has('save_review')) reviewManagement::reviewRegister();
-
+ if (Input::has('tocontact')) contactManagement::contactform();
