+++ review.php
@@ -14,9 +14,12 @@ include('includes/autoloader.php');
    <div class="row"> 
   <div class="col-md-4"><!--Linker kant--></div>
     <div class="col-md-4">
+        
+        
 <form method="post" accept-charset="utf-8">
-   
-    	<h3>Review This Product</h3>	
+        <?php if($_SESSION['user']['role'] == 'Ouder' ){        
+      	 echo '         
+    	<h3>Schrijf hier uw review</h3>	
     	<!-- Rating -->
     <p>
     	<label for="rating">Rating</label>
@@ -35,6 +38,8 @@ include('includes/autoloader.php');
     	<input type="submit" value="Plaats Review" name="save_review">
 	
 </form>
+            ';}
+        ?>
     <div class="col-md-4"><!--Rechter kant--></div> 
     </div>
